Authentication (#5)
* add basic signup with bcrypt-hashed password

* add login and signup with JWT

* refactor and add /me test endpoint

* minor refactor

* refactor and move functions

* minor refactor

+++ lambdas/routes/users/main.py
@@ -2,40 +2,21 @@ from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
+from lambdas.routes.authentication.main import router as authentication_router
+from ..authentication.utils import check_logged_in
 from ...db.main import get_db_session, User
-from ...dtos.users import CreateRequestDTO, CreateResponseDTO, UpdateRequestDTO
+from ...dtos.users import UpdateRequestDTO
 
 router = APIRouter(prefix="/users", tags=["users"])
-
-
-@router.post('', response_model=CreateResponseDTO)
-async def create(user: CreateRequestDTO, db: Session = Depends(get_db_session)):
-    new_user = User(name=user.name, email=user.email, username=user.username)
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user, attribute_names=['id', 'name', 'email', 'username'])
-    return {
-        'message': 'User created successfully',
-        'user': {
-            'id': str(new_user.id),
-            'name': new_user.name,
-            'email': new_user.email,
-            'username': new_user.username
-        }
-    }
-
-
-@router.get('/{user_id}')
-async def get_user(user_id: str, db: Session = Depends(get_db_session)):
-    user = db.query(User).filter(User.id == user_id).first()
-    if user is None:
-        raise HTTPException(status_code=404, detail='User not found')
-
-    return {'user': user}
+router.include_router(authentication_router)
 
 
 @router.patch('/{user_id}')
-async def update(user_id: str, user: UpdateRequestDTO, db: Session = Depends(get_db_session)):
+async def update(user_id: str, user: UpdateRequestDTO, db: Session = Depends(get_db_session),
+                 is_logged_in: bool = Depends(check_logged_in)):
+    if not is_logged_in:
+        raise HTTPException(status_code=401, detail='Not logged in', headers={'WWW-Authenticate': 'Bearer'})
+
     fields = extract_set_fields(user)
     if len(fields) == 0:
         raise HTTPException(status_code=400, detail='Body cannot be empty')
@@ -49,7 +30,10 @@ async def update(user_id: str, user: UpdateRequestDTO, db: Session = Depends(get
 
 
 @router.delete('/{user_id}')
-async def delete(user_id: str, db: Session = Depends(get_db_session)):
+async def delete(user_id: str, db: Session = Depends(get_db_session), is_logged_in: bool = Depends(check_logged_in)):
+    if not is_logged_in:
+        raise HTTPException(status_code=401, detail='Not logged in', headers={'WWW-Authenticate': 'Bearer'})
+    
     count = db.query(User).filter(User.id == user_id).delete()
     if count == 0:
         raise HTTPException(status_code=404, detail='User not found')
